+++ backend/src/model/config/web_ui.rs
@@ -6,6 +6,7 @@ use crate::model::{macros, WebAuthConfig};
 pub struct WebUiConfig {
     pub enabled: bool,
     pub user_ui_enabled: bool,
+    pub content_security_policy: bool,
     pub path: Option<String>,
     pub auth: Option<WebAuthConfig>,
     pub player_server: Option<String>,
@@ -30,6 +31,7 @@ impl From<&WebUiConfigDto> for WebUiConfig {
         Self {
             enabled: dto.enabled,
             user_ui_enabled: dto.user_ui_enabled,
+            content_security_policy: dto.content_security_policy,
             path: dto.path.clone(),
             auth: dto.auth.as_ref().map(Into::into),
             player_server: dto.player_server.clone(),
@@ -41,6 +43,7 @@ impl From<&WebUiConfig> for WebUiConfigDto {
         Self {
             enabled: instance.enabled,
             user_ui_enabled: instance.user_ui_enabled,
+            content_security_policy: instance.content_security_policy,
             path: instance.path.clone(),
             auth: instance.auth.as_ref().map(Into::into),
             player_server: instance.player_server.clone(),