+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - EPG Config View
 - Fixed loading users for WebUI from user DB
 - Fixed auto EPG for batch inputs
-- ContentSecurityPolicy can be enabled
+- Content Security Policies configurable via config
 
 # 3.1.5 (2025-08-14)
 - Hot reload for config